+++ app/lib/package/backend.dart
@@ -29,6 +29,7 @@ import '../job/backend.dart';
 import '../publisher/backend.dart';
 import '../service/email/backend.dart';
 import '../service/email/models.dart';
+import '../service/rate_limit/rate_limit.dart';
 import '../service/secret/backend.dart';
 import '../shared/configuration.dart';
 import '../shared/datastore.dart';
@@ -1046,6 +1047,12 @@ class PackageBackend {
       throw AssertionError(
           'Package "${newVersion.package}" has no admin email to notify.');
     }
+
+    // check rate limits before the transaction
+    await verifyPackageUploadRateLimit(
+      agent: agent,
+      package: newVersion.package,
+    );
 
     final email = createPackageUploadedEmail(
       packageName: newVersion.package,