+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Note: Started to rate limit package uploads using (locally cached) AuditLogRecords.
 
 ## `20230516t150900-all`
  * Bumped runtimeVersion to `2023.05.16`.